// shared/schema.d.ts
export declare const insertProjectApiSchema: z.ZodObject<{
    title: z.ZodString;
    description: z.ZodString;
    techStack: z.ZodDefault<z.ZodArray<z.ZodString, "many">>;
    imageUrl: z.ZodString;
    githubUrl: z.ZodOptional<z.ZodNullable<z.ZodEffects<z.ZodString, string, string>>>;
    liveUrl: z.ZodOptional<z.ZodNullable<z.ZodEffects<z.ZodString, string, string>>>;
    category: z.ZodString;
    problemStatement: z.ZodOptional<z.ZodDefault<z.ZodNullable<z.ZodString>>>;
    motivation: z.ZodOptional<z.ZodDefault<z.ZodNullable<z.ZodString>>>;
    systemDesign: z.ZodOptional<z.ZodDefault<z.ZodNullable<z.ZodString>>>;
    challenges: z.ZodOptional<z.ZodDefault<z.ZodNullable<z.ZodString>>>;
    learnings: z.ZodOptional<z.ZodDefault<z.ZodNullable<z.ZodString>>>;
}, "strip", z.ZodTypeAny, {
    title: string;
    description: string;
    techStack: string[];
    imageUrl: string;
    category: string;
    githubUrl?: string | null | undefined;
    liveUrl?: string | null | undefined;
    problemStatement?: string | null | undefined;
    motivation?: string | null | undefined;
    systemDesign?: string | null | undefined;
    challenges?: string | null | undefined;
    learnings?: string | null | undefined;
}, {
    title: string;
    description: string;
    imageUrl: string;
    category: string;
    techStack?: string[] | undefined;
    githubUrl?: string | null | undefined;
    liveUrl?: string | null | undefined;
    problemStatement?: string | null | undefined;
    motivation?: string | null | undefined;
    systemDesign?: string | null | undefined;
    challenges?: string | null | undefined;
    learnings?: string | null | undefined;
}>;
export declare const skillSchema: z.ZodObject<{
    id: z.ZodNumber;
    name: z.ZodString;
    category: z.ZodString;
    icon: z.ZodString;
}, "strip", z.ZodTypeAny, {
    id: number;
    name: string;
    category: string;
    icon: string;
}, {
    id: number;
    name: string;
    category: string;
    icon: string;
}>;
export declare const insertSkillApiSchema: z.ZodObject<{
    name: z.ZodString;
    category: z.ZodString;
    icon: z.ZodDefault<z.ZodString>;
}, "strip", z.ZodTypeAny, {
    name: string;
    category: string;
    icon: string;
}, {
    name: string;
    category: string;
    icon?: string | undefined;
}>;
// ...
export declare const insertExperienceApiSchema: z.ZodObject<{
    role: z.ZodString;
    organization: z.ZodString;
    period: z.ZodString;
    description: z.ZodString;
    type: z.ZodDefault<z.ZodString>;
}, "strip", z.ZodTypeAny, {
    description: string;
    role: string;
    organization: string;
    period: string;
    type: string;
}, {
    description: string;
    role: string;
    organization: string;
    period: string;
    type?: string | undefined;
}>;
// ...
export declare const insertMessageApiSchema: z.ZodObject<{
    name: z.ZodString;
    email: z.ZodString;
    subject: z.ZodDefault<z.ZodString>;
    message: z.ZodString;
}, "strip", z.ZodTypeAny, {
    name: string;
    email: string;
    subject: string;
    message: string;
}, {
    name: string;
    email: string;
    message: string;
    subject?: string | undefined;
}>;